+++ Product.h
@@ -10,7 +10,7 @@ class Product
 {
 public:
     Product(string name, skills::skillsType skill, int materialCost,
-            int timeCost, string materialNeeded, int value);
+            int timeCost, string materialNeeded, double value);
 
     string getName();
     skills::skillsType getSkillType();
@@ -18,7 +18,7 @@ public:
     int getMaterialCost();
     int getTimeCost();
 
-    int getValue();
+    double getValue();
 
 private:
 
@@ -28,7 +28,7 @@ private:
     int materialCost;
     int timeCost;
 
-    int value;
+    double value;
     //TO DO:
     //          - Time cost should be standard
     //          - it could be 3 minutes but this would be with one worker, with stat 1/100
